% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_run_query_parallel.R
\name{create_path_url}
\alias{create_path_url}
\title{Create path to a firestore collection (event core responses)}
\usage{
create_path_url(partial_path, collection_id, project_id)
}
\arguments{
\item{partial_path}{Initial part of path to the collection.}

\item{collection_id}{Event core responses collection.}

\item{project_id}{Firebase project ID.}
}
\value{
Character.
}
\description{
Create path to a firestore collection (event core responses)
}
