% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_run_query_parallel.R
\name{create_query_body}
\alias{create_query_body}
\title{Create a jsonifed query body for a request}
\usage{
create_query_body(
  selected_fields,
  filter_by,
  operation,
  value_type,
  value,
  collection_id
)
}
\arguments{
\item{selected_fields}{Fields you want to return in response.}

\item{filter_by}{Fields to filter by.}

\item{operation}{Type of operation.}

\item{value_type}{Type of value.}

\item{value}{Value.}

\item{collection_id}{ID of firestore collection.}
}
\value{
Character.
}
\description{
Create a jsonifed query body for a request
}
