% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_run_query_parallel.R
\name{create_request}
\alias{create_request}
\title{Create an HTTP request for firestore.}
\usage{
create_request(base_url, path_url, id_token, structured_query)
}
\arguments{
\item{base_url}{Base URL.}

\item{path_url}{URL of the path to a collection.}

\item{id_token}{Access Token.}

\item{structured_query}{Query parameters in JSON format.}
}
\value{
Request.
}
\description{
Create an HTTP request for firestore.
}
