% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_run_query_parallel.R
\name{create_select_clause}
\alias{create_select_clause}
\title{Jsonify the selected fields in query}
\usage{
create_select_clause(selected_fields)
}
\arguments{
\item{selected_fields}{Fields you want to return in response.}
}
\value{
Character.
}
\description{
Jsonify the selected fields in query
}
