% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_base_url.R
\name{frstore_base_url}
\alias{frstore_base_url}
\title{Get the base url for Cloud Firestore}
\usage{
frstore_base_url()
}
\value{
Character. Base url.
}
\description{
Get the base url for Cloud Firestore
}
\examples{
\dontrun{
frstore_base_url()
}
}
