% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_create_document.R
\name{frstore_create_document}
\alias{frstore_create_document}
\title{Perform POST request to Cloud Firestore database to post data}
\usage{
frstore_create_document(document_path, id_token, data = NULL)
}
\arguments{
\item{document_path}{Character. The path to a collection or a document.}

\item{id_token}{Character. Firebase authentication token.}

\item{data}{List with a specific structure. Defaults to \link{NULL}. See Details.}
}
\value{
Request.
}
\description{
Perform POST request to Cloud Firestore database to post data
}
\details{
Visit \href{https://cloud.google.com/firestore/docs/reference/rest/v1beta1/projects.databases.documents/createDocument}{createDocument page of Firestore REST API docs}
for more details. The \code{data} must be in this format:
\preformatted{
list(
  fields = list(
    field1 = list("<someValue>" = <number>),
    field2 = list("<someValue>" = <character>)
  )
)
}
Where \code{someValue} can be:
\itemize{
\item \code{arrayValue}
\item \code{booleanValue}
\item \code{bytesValue}
\item \code{doubleValue}
\item \code{geoPointValue}
\item \code{integerValue}
\item \code{mapValue}
\item \code{nullValue},
\item \code{referenceValue}
\item \code{stringValue}
\item \code{timestampValue}
}

We have tested only \code{arrayValue}, \code{doubleValue},
\code{integerValue}, and \code{stringValue}.
}
\examples{
\dontrun{
library(frbs)
# Sign up via Firebase authentication:
frbs_sign_up(email = "<EMAIL>", password = "<PASSWORD>")
# Sign in:
foo <- frbs_sign_in(email = "<EMAIL>", password = "<PASSWORD>")
# Create a document without specifying data:
frstore_create_document("test/firstDoc", foo$idToken)
# Create a document in a subcollection with data:
data_list <- list(
  fields = list(
    age = list("integerValue" = 36),
    name = list("stringValue" = "merry")
 )
)
frstore_create_document("test/firstDoc/firstCollection/doc", foo$idToken, data_list)
# Create a document with data:
frstore_create_document("test/secondDoc", foo$idToken, data_list)
}
}
