% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_delete.R
\name{frstore_delete}
\alias{frstore_delete}
\title{Perform DELETE request to Cloud Firestore database to delete data}
\usage{
frstore_delete(document_path, id_token)
}
\arguments{
\item{document_path}{Character. The path to a collection or a document.}

\item{id_token}{Character. Firebase authentication token.}
}
\value{
Request.
}
\description{
Perform DELETE request to Cloud Firestore database to delete data
}
\details{
Visit \href{https://cloud.google.com/firestore/docs/reference/rest/v1beta1/projects.databases.documents/delete}{delete page of Firestore REST API docs}
for more details.
}
\examples{
\dontrun{
library(frbs)
# Sign up via Firebase authentication:
frbs_sign_up(email = "<EMAIL>", password = "<PASSWORD>")
# Sign in:
foo <- frbs_sign_in(email = "<EMAIL>", password = "<PASSWORD>")
# Suppose there is an existing document at
# test/firstDoc/firstCollection/doc
# and we want to delete it:
frstore_delete("test/firstDoc/firstCollection/doc", foo$idToken)
}
}
