% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_error_handler.R
\name{frstore_error_handler}
\alias{frstore_error_handler}
\title{Error handler}
\usage{
frstore_error_handler(e)
}
\arguments{
\item{e}{Error.}
}
\description{
Used in the \code{tryCatch()} blocks of \verb{frstore_*()} functions
}
