% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_get.R
\name{frstore_get}
\alias{frstore_get}
\title{Perform GET request to Cloud Firestore database to get data}
\usage{
frstore_get(document_path, id_token, fields = NULL)
}
\arguments{
\item{document_path}{Character. The path to a collection or a document.}

\item{id_token}{Character. Firebase authentication token.}

\item{fields}{Character. One or more fields to return. Defaults to \link{NULL} for all fields.}
}
\value{
A list with requested document(s).
}
\description{
Perform GET request to Cloud Firestore database to get data
}
\details{
Visit \href{https://cloud.google.com/firestore/docs/reference/rest/v1beta1/projects.databases.documents/get}{get page of Firestore REST API docs}
for more details.
}
\examples{
\dontrun{
# Suppose there is a collection named test with a document named doc.
library(frbs)
# Sign up via Firebase authentication:
frbs_sign_up(email = "<EMAIL>", password = "<PASSWORD>")
# Sign in:
foo <- frbs_sign_in(email = "<EMAIL>", password = "<PASSWORD>")
# Get document(s):
frstore_get("test", foo$idToken)
frstore_get("test/doc", foo$idToken)
frstore_get("test/doc", foo$idToken, fields = c("age"))
}
}
