% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_patch.R
\name{frstore_patch}
\alias{frstore_patch}
\title{Perform PATCH request to Cloud Firestore database to update data}
\usage{
frstore_patch(document_path, id_token, data)
}
\arguments{
\item{document_path}{Character. The path to a collection or a document.}

\item{id_token}{Character. Firebase authentication token.}

\item{data}{List with a specific structure. See Details.}
}
\value{
Request.
}
\description{
Perform PATCH request to Cloud Firestore database to update data
}
\details{
Visit \href{https://cloud.google.com/firestore/docs/reference/rest/v1beta1/projects.databases.documents/patch}{patch page of Firestore REST API docs}
for more details. The \code{data} must be in this format:
\preformatted{
list(
  fields = list(
    field1 = list("<someValue>" = <number>),
    field2 = list("<someValue>" = <character>)
  )
)
}
Where \code{someValue} can be:
\itemize{
\item \code{arrayValue}
\item \code{booleanValue}
\item \code{bytesValue}
\item \code{doubleValue}
\item \code{geoPointValue}
\item \code{integerValue}
\item \code{mapValue}
\item \code{nullValue},
\item \code{referenceValue}
\item \code{stringValue}
\item \code{timestampValue}
}

We have tested only \code{arrayValue}, \code{doubleValue},
\code{integerValue}, and \code{stringValue}.
}
\examples{
\dontrun{
library(frbs)
# Sign up via Firebase authentication:
frbs_sign_up(email = "<EMAIL>", password = "<PASSWORD>")
# Sign in:
foo <- frbs_sign_in(email = "<EMAIL>", password = "<PASSWORD>")
# Suppose there is an existing document at
# test/firstDoc/firstCollection/doc
# and we want to update it with new data:
data_list <- list(
   fields = list(
     age = list("integerValue" = 3600),
     name = list("stringValue" = "merryyyy")
   )
)
frstore_patch("test/firstDoc/firstCollection/doc", foo$idToken, data_list)
}
}
