% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_project_id.R
\name{frstore_project_id}
\alias{frstore_project_id}
\title{Get Cloud Firebase Project ID}
\usage{
frstore_project_id()
}
\value{
Character. Firebase Project ID.
}
\description{
Get Cloud Firebase Project ID
}
\examples{
\dontrun{
frstore_project_id()
}
}
