% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_req.R
\name{frstore_req}
\alias{frstore_req}
\title{Create basic structure of Cloud Firestore request without performing it}
\usage{
frstore_req(
  document_path,
  id_token,
  project_id = frstore_project_id(),
  base_url = frstore_base_url()
)
}
\arguments{
\item{document_path}{Character. The path to a collection or a document.}

\item{id_token}{Character. Firebase authentication token.}

\item{project_id}{Character. Firebase project ID. Defaults to \code{\link[=frstore_project_id]{frstore_project_id()}}.}

\item{base_url}{Character. Cloud Firestore base url. Defaults to \code{\link[=frstore_base_url]{frstore_base_url()}}.}
}
\value{
An unperformed request. Contains the full path to a collection (or a document) with request headers.
}
\description{
Other functions add steps to \code{frstore_req()} and then perform the request.
}
