% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_run_query.R
\name{frstore_run_query}
\alias{frstore_run_query}
\title{Run a query to get specific data from Cloud Firestore.}
\usage{
frstore_run_query(
  collection_path,
  id_token,
  field,
  operation,
  value_type,
  value,
  selected_fields = NULL,
  project_id = frstore_project_id(),
  base_url = frstore_base_url()
)
}
\arguments{
\item{collection_path}{Character. The path to a collection or a subcollection that contains the document(s) you want to query.}

\item{id_token}{Character. Firebase authentication token.}

\item{field}{Character. \code{fieldPath} in \code{structuredQuery}. See Details.}

\item{operation}{Character. \code{op} in \code{structuredQuery}. See Details.}

\item{value_type}{Character. Part of \code{value} in \code{structuredQuery}. See Details.}

\item{value}{Character. Part of \code{value} in \code{structuredQuery}. See Details.}

\item{selected_fields}{Character. Part of \code{select} in \code{structuredQuery}. See Details.}

\item{project_id}{Character. Firebase project ID. Defaults to \code{\link[=frstore_project_id]{frstore_project_id()}}.}

\item{base_url}{Character. Cloud Firestore base url. Defaults to \code{\link[=frstore_base_url]{frstore_base_url()}}.}
}
\value{
A list with requested document(s).
}
\description{
Run a query to get specific data from Cloud Firestore.
}
\details{
Visit \href{https://cloud.google.com/firestore/docs/reference/rest/v1beta1/projects.databases.documents/runQuery}{runQuery page of Firestore REST API docs}
for more details.

\code{structuredQuery} contains these parameters:
\preformatted{
 {
 "structuredQuery": {
   <<select_clause>>
   "where": {
     "fieldFilter": {
       "field": {
         "fieldPath": "<<field>>"
       },
       "op": "<<operation>>",
       "value": {
         "<<value_type>>": "<<value>>"
       }
     }
   },
   "from": [
     {
       "collectionId": "<<collection>>"
     }
   ]
  }
 }
}
Each value in \verb{<<>>} corresponds to a parameter. \code{op} can be \code{LESS_THAN}, \code{LESS_THAN_OR_EQUAL},
\code{GREATER_THAN}, \code{GREATER_THAN_OR_EQUAL}, \code{EQUAL}, \code{NOT_EQUAL}, \code{ARRAY_CONTAINS}, \code{IN},
\code{ARRAY_CONTAINS_ANY}, and \code{NOT_IN}.
}
\examples{
\dontrun{
library(frbs)
# Sign up via Firebase authentication:
frbs_sign_up(email = "<EMAIL>", password = "<PASSWORD>")
# Sign in:
foo <- frbs_sign_in(email = "<EMAIL>", password = "<PASSWORD>")
# Suppose there is an existing subcollection at
# test/firstDoc/firstCollection and
# we want to get all docs where name matches "merry":
frstore_run_query("test/firstDoc/firstCollection",
  foo$idToken,
  field = "name",
  operation = "EQUAL",
  value_type = "stringValue",
  value = "merry")
}
}
