% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frstore_run_query_parallel.R
\name{frstore_run_query_sequential}
\alias{frstore_run_query_sequential}
\title{Sequential version of frstore_run_query}
\usage{
frstore_run_query_sequential(
  query_list,
  id_token,
  project_id = frstore_project_id(),
  base_url = frstore_base_url()
)
}
\arguments{
\item{query_list}{List of lists of query parameters for each request.}

\item{id_token}{Access Token.}

\item{project_id}{Firebase project ID.}

\item{base_url}{Base URL.}
}
\value{
List
}
\description{
Sequential version of frstore_run_query
}
